requirements.txt
# Risk Monitoring System v2.2 - Dependencies
# pip install -r requirements.txt

# FastAPI & ASGI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Neo4j
neo4j>=5.0.0
langchain-neo4j>=0.1.0

# OpenAI
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# Validation
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0

# Logging
loguru>=0.7.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
